Fixes FCollada build warnings with GCC, based on patch by Markus, fixes #1851
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@13629 3db68df2-c116-0410-a063-a993310a9797

+++ libraries/source/fcollada/src/FCollada/FUtils/FUDebug.cpp
@@ -70,7 +70,7 @@ void FUDebug::DebugOut(uint8 verbosity, const char* message, ...)
 void FUDebug::DebugOutV(uint8 verbosity, const char* filename, uint32 line, const char* message, va_list& vars)
 {
 	char buffer[256];
-	snprintf(buffer, 256, "[%s@%lu] ", filename, line);
+	snprintf(buffer, 256, "[%s@%lu] ", filename, (unsigned long)line);
 	buffer[255] = 0;
 	DebugString(buffer);
